Move deleteOldTasks and setTasksArray to logic.js
